fixing long warnings

+++ Sprocket/Libraries/SWRevealViewController/SWRevealViewController.m
@@ -1347,7 +1347,7 @@ const int FrontViewPositionNone = 0xff;
     
     NSTimeInterval duration = animated?_toggleAnimationDuration:0.0;
     NSTimeInterval firstDuration = duration;
-    int initialPosDif = labs( _frontViewPosition - preReplacementPosition );
+    long initialPosDif = labs( _frontViewPosition - preReplacementPosition );
     if ( initialPosDif == 1 ) firstDuration *= 0.8;
     else if ( initialPosDif == 0 ) firstDuration = 0;
